+++ sonarqube/run.sh
@@ -6,12 +6,14 @@ if [ "${1:0:1}" != '-' ]; then
   exec "$@"
 fi
 
-# Update sonar.properties for crowd plugin
-echo "sonar.security.realm=Crowd" >> conf/sonar.properties
-echo "crowd.url=$SONARQUBE_CROWD_URL" >> conf/sonar.properties
-echo "crowd.application=$SONARQUBE_CROWD_APP" >> conf/sonar.properties
-echo "crowd.password=$SONARQUBE_CROWD_PWD" >> conf/sonar.properties
-echo "sonar.security.localUsers=admin" >> conf/sonar.properties
+if [ $SONAR_AUTH_CROWD == true ]; then
+  # Update sonar.properties for crowd plugin
+  echo "sonar.security.realm=Crowd" >> conf/sonar.properties
+  echo "crowd.url=$SONARQUBE_CROWD_URL" >> conf/sonar.properties
+  echo "crowd.application=$SONARQUBE_CROWD_APP" >> conf/sonar.properties
+  echo "crowd.password=$SONARQUBE_CROWD_PWD" >> conf/sonar.properties
+  echo "sonar.security.localUsers=admin" >> conf/sonar.properties
+fi
 
 # upgrade to 7.3
 rm $SONARQUBE_HOME/extensions/plugins/*.jar || true